requirements.txt
langdetect
python-dotenv
langchain
langchain-community
faiss-cpu
langchain-google-genai
crewai
crewai-tools
